[fix/#18] RepositoryService 사용
