+++ src/main/java/com/clokey/server/domain/category/application/CategoryRepositoryService.java
@@ -1,6 +1,6 @@
 package com.clokey.server.domain.category.application;
 
-import com.clokey.server.domain.category.dto.CategoryResponseDTO;
+import com.clokey.server.domain.model.Category;
 
 import java.util.List;
 
@@ -8,5 +8,5 @@ public interface CategoryRepositoryService {
 
     boolean categoryExist(Long categoryId);
 
-    List<CategoryResponseDTO> getAllCategories();
+    List<Category> getAllCategories();
 }
